tasklist-15-kadai1

+++ app/Http/Controllers/TasksController.php
@@ -12,22 +12,38 @@ class TasksController extends Controller
     public function index()
     {
           
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) { 
+            
+            $user = \Auth::user();
+            
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+        }
 
+            return view('welcome', $data);
     }
+
 
     public function create()
     {
         $task = new Task;
 
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.create', [
             'task' => $task,
         ]);
+        }
     }
+
 
     public function store(Request $request)
     {
@@ -36,34 +52,40 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
 
-        
-        return redirect('/');
+        return back();
     }
+
 
     public function show($id)
     {
-        
         $task = Task::findOrFail($id);
 
+        if (\Auth::id() === $task->user_id) {
         
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }
     }
+
 
     public function edit($id)
     {
-        
         $task = Task::findOrFail($id);
-        return view('tasks.edit', [
+        
+        if (\Auth::id() === $task->user_id) {
+           
+            return view('tasks.edit', [
             'task' => $task,
-        ]);
+            ]);
+        }
     }
+   
    
     public function update(Request $request, $id)
     {
@@ -73,19 +95,25 @@ class TasksController extends Controller
         ]);
         
         $task = Task::findOrFail($id);
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-       
+        
+        if (\Auth::id() === $task->user_id) {
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
+        }
+        
         return redirect('/');
     }
 
    
     public function destroy($id)
     {
-        
         $task = Task::findOrFail($id);
-        $task->delete();
+
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
         return redirect('/');
     }
 }
